requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
jinja2==3.1.2
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2